chore: more reorg
