+++ lib/index.js
@@ -1,8 +1,6 @@
-const { mkdirSync, writeFileSync } = require('fs');
-const { dirname, join, relative, resolve } = require('path');
+const { relative, resolve } = require('path');
 
-const { standardizeFilePaths } = require('./helpers');
-const { beforeRunHook, getCompilerHooks, moduleAssetHook } = require('./hooks');
+const { beforeRunHook, emitHook, getCompilerHooks, moduleAssetHook } = require('./hooks');
 
 const emitCountMap = new Map();
 
@@ -18,6 +16,7 @@ const defaults = {
     return JSON.stringify(manifest, null, 2);
   },
   sort: null,
+  transform: (files) => files,
   transformExtensions: /^(gz|map)$/i,
   writeToFileEmit: false
 };
@@ -27,180 +26,24 @@ class WebpackManifestPlugin {
     this.options = Object.assign({}, defaults, opts);
   }
 
-  getFileType(str) {
-    const replaced = str.replace(/\?.*/, '');
-    const split = replaced.split('.');
-    let ext = split.pop();
-    if (this.options.transformExtensions.test(ext)) {
-      ext = `${split.pop()}.${ext}`;
-    }
-    return ext;
-  }
-
   apply(compiler) {
     const moduleAssets = {};
     const manifestFileName = resolve(compiler.options.output.path, this.options.fileName);
     const manifestAssetId = relative(compiler.options.output.path, manifestFileName);
     const beforeRun = beforeRunHook.bind(this, { emitCountMap, manifestFileName });
+    const emit = emitHook.bind(this, {
+      compiler,
+      emitCountMap,
+      manifestAssetId,
+      manifestFileName,
+      moduleAssets,
+      options: this.options
+    });
     const moduleAsset = moduleAssetHook.bind(this, { moduleAssets });
     const hookOptions = {
       name: 'WebpackManifestPlugin',
       stage: Infinity
     };
-
-    const emit = function emit(compilation) {
-      const emitCount = emitCountMap.get(manifestFileName) - 1;
-      emitCountMap.set(manifestFileName, emitCount);
-
-      const seed = this.options.seed || {};
-      const publicPath =
-        this.options.publicPath !== null
-          ? this.options.publicPath
-          : compilation.options.output.publicPath;
-      const stats = compilation.getStats().toJson({
-        // Disable data generation of everything we don't use
-        all: false,
-        // Add asset Information
-        assets: true,
-        // Show cached assets (setting this to `false` only shows emitted files)
-        cachedAssets: true
-      });
-
-      let files = compilation.chunks.reduce(
-        (files, chunk) =>
-          chunk.files.reduce((files, path) => {
-            let name = chunk.name ? chunk.name : null;
-
-            if (name) {
-              name = `${name}.${this.getFileType(path)}`;
-            } else {
-              // For nameless chunks, just map the files directly.
-              name = path;
-            }
-
-            return files.concat({
-              path,
-              chunk,
-              name,
-              isInitial: chunk.isOnlyInitial(),
-              isChunk: true,
-              isAsset: false,
-              isModuleAsset: false
-            });
-          }, files),
-        []
-      );
-
-      // module assets don't show up in assetsByChunkName.
-      // we're getting them this way;
-      files = stats.assets.reduce((files, asset) => {
-        const name = moduleAssets[asset.name];
-        if (name) {
-          return files.concat({
-            path: asset.name,
-            name,
-            isInitial: false,
-            isChunk: false,
-            isAsset: true,
-            isModuleAsset: true
-          });
-        }
-
-        const isEntryAsset = asset.chunks.length > 0;
-        if (isEntryAsset) {
-          return files;
-        }
-
-        return files.concat({
-          path: asset.name,
-          name: asset.name,
-          isInitial: false,
-          isChunk: false,
-          isAsset: true,
-          isModuleAsset: false
-        });
-      }, files);
-
-      files = files.filter((file) => {
-        // Don't add hot updates to manifest
-        const isUpdateChunk = file.path.indexOf('hot-update') >= 0;
-        // Don't add manifest from another instance
-        const isManifest =
-          typeof emitCountMap.get(join(compiler.options.output.path, file.name)) !== 'undefined';
-
-        return !isUpdateChunk && !isManifest;
-      });
-
-      // Append optional basepath onto all references.
-      // This allows output path to be reflected in the manifest.
-      if (this.options.basePath) {
-        files = files.map((file) => {
-          file.name = this.options.basePath + file.name;
-          return file;
-        });
-      }
-
-      if (publicPath) {
-        // Similar to basePath but only affects the value (similar to how
-        // output.publicPath turns require('foo/bar') into '/public/foo/bar', see
-        // https://github.com/webpack/docs/wiki/configuration#outputpublicpath
-        files = files.map((file) => {
-          file.path = publicPath + file.path;
-          return file;
-        });
-      }
-
-      files = files.map(standardizeFilePaths);
-
-      if (this.options.filter) {
-        files = files.filter(this.options.filter);
-      }
-
-      if (this.options.map) {
-        files = files.map(this.options.map).map(standardizeFilePaths);
-      }
-
-      if (this.options.sort) {
-        files = files.sort(this.options.sort);
-      }
-
-      let manifest;
-      if (this.options.generate) {
-        const entrypointsArray = Array.from(compilation.entrypoints.entries());
-        const entrypoints = entrypointsArray.reduce(
-          (e, [name, entrypoint]) => Object.assign(e, { [name]: entrypoint.getFiles() }),
-          {}
-        );
-        manifest = this.options.generate(seed, files, entrypoints);
-      } else {
-        manifest = files.reduce((manifest, file) => {
-          manifest[file.name] = file.path;
-          return manifest;
-        }, seed);
-      }
-
-      const isLastEmit = emitCount === 0;
-      if (isLastEmit) {
-        const output = this.options.serialize(manifest);
-
-        compilation.assets[manifestAssetId] = {
-          source() {
-            return output;
-          },
-          size() {
-            return output.length;
-          }
-        };
-
-        if (this.options.writeToFileEmit) {
-          mkdirSync(dirname(manifestFileName), { recursive: true });
-          writeFileSync(manifestFileName, output);
-          // fse.outputFileSync(manifestFileName, output);
-        }
-      }
-
-      getCompilerHooks(compiler).afterEmit.call(manifest);
-    }.bind(this);
 
     compiler.hooks.compilation.tap(hookOptions, (compilation) => {
       compilation.hooks.moduleAsset.tap(hookOptions, moduleAsset);
